changes from the group...
------------------------------------------------------------------------
r10738 | jonas.larsen | 2010-03-15 06:13:25 -0600 (Mon, 15 Mar 2010) | 11 lines
Changed paths:
  M /active/code/create.py
  M /active/code/flagging/CMakeLists.txt
  M /active/code/install/target.cmake


      New Development: No
           JIRA Issue: CAS-1936
     Ready to Release: Yes
    Interface Changes: No
What Interface Changed: No
        Test Programs: No
 Put in Release Notes: No
            Module(s): 
          Description: Do not use assay for flagging unit tests.

------------------------------------------------------------------------
r10739 | jonas.larsen | 2010-03-15 07:09:53 -0600 (Mon, 15 Mar 2010) | 11 lines
Changed paths:
  M /active/code/xmlcasa/scripts/regressions/admin/imageTest.py


      New Development: No
           JIRA Issue: No
     Ready to Release: Yes
    Interface Changes: No
What Interface Changed: No
        Test Programs: No
 Put in Release Notes: No
            Module(s): 
          Description: Have to go back to im.fitsky() until I have time to figure out how to use im.fitcomponents() without getting an exception from the following resid=pylab.array(a['pixels']).

------------------------------------------------------------------------
r10740 | jonas.larsen | 2010-03-15 07:24:23 -0600 (Mon, 15 Mar 2010) | 11 lines
Changed paths:
  M /active/code/flagging/implement/Flagging/Flagger.cc


      New Development: No
           JIRA Issue: CAS-2021
     Ready to Release: Yes
    Interface Changes: No
What Interface Changed: No
        Test Programs: No
 Put in Release Notes: No
            Module(s): 
          Description: Initializing the object in the constructor is a great idea.

------------------------------------------------------------------------
r10741 | jonas.larsen | 2010-03-15 07:45:17 -0600 (Mon, 15 Mar 2010) | 11 lines
Changed paths:
  M /active/code/flagging/implement/Flagging/RFASelector2.cc


      New Development: No
           JIRA Issue: CAS-2021
     Ready to Release: Yes
    Interface Changes: No
What Interface Changed: No
        Test Programs: No
 Put in Release Notes: No
            Module(s): 
          Description: Input parameter validation... (Evidently, this code was written before the invention of NULL, hence they invented the "null" MS (or that's the most conceivable explanation).)

------------------------------------------------------------------------
r10742 | darrell.schiebel | 2010-03-15 10:26:56 -0600 (Mon, 15 Mar 2010) | 11 lines
Changed paths:
  M /active/code/install/makedefs
  M /active/code/xmlcasa/apps/module_paramgui/mksetupdotpy.sh
  M /active/code/xmlcasa/install/mksetupdotpy.sh

      New Development: no
           JIRA Issue: no
     Ready to Release: yes
    Interface Changes: none
What Interface Changed: none
        Test Programs: none
 Put in Release Notes: no

          Description: consolidate ccmtools and make it easier to build (no source changes, just consolidation)


------------------------------------------------------------------------
r10743 | jonas.larsen | 2010-03-15 11:00:27 -0600 (Mon, 15 Mar 2010) | 11 lines
Changed paths:
  M /active/code/CMakeLists.txt


      New Development: No
           JIRA Issue: CAS-1936
     Ready to Release: Yes
    Interface Changes: No
What Interface Changed: No
        Test Programs: No
 Put in Release Notes: No
            Module(s): 
          Description: Oops, the COREINCD definition for ASAP had broken when switching to casa_find() for finding CASACore.

------------------------------------------------------------------------
r10744 | jonas.larsen | 2010-03-15 12:00:43 -0600 (Mon, 15 Mar 2010) | 11 lines
Changed paths:
  M /active/code/install/target.cmake


      New Development: No
           JIRA Issue: CAS-1936
     Ready to Release: Yes
    Interface Changes: No
What Interface Changed: No
        Test Programs: No
 Put in Release Notes: No
            Module(s): 
          Description: There are no explicit dl libraries on Mac.

------------------------------------------------------------------------
r10745 | remy.indebetouw | 2010-03-15 14:29:27 -0600 (Mon, 15 Mar 2010) | 10 lines
Changed paths:
  M /active/code/xmlcasa/scripts/task_simdata.py

      New Development: no
           JIRA Issue: no
     Ready to Release: yes
    Interface Changes: no
        Test Programs: none
 Put in Release Notes: no

          Description: bugfix imsize param type list failure


------------------------------------------------------------------------
r10746 | dave.mehringer | 2010-03-15 14:38:32 -0600 (Mon, 15 Mar 2010) | 14 lines
Changed paths:
  M /active/casacore/images/Images/ImageAnalysis.cc


      New Development: No
           JIRA Issue: No
     Ready to Release: Yes
    Interface Changes: No
What Interface Changed: Please list interface changes
        Test Programs: List test programs
 Put in Release Notes: Yes/No
            Module(s): Module Names change impacts.
          Description: added a LogOrigin call 

	


------------------------------------------------------------------------
r10747 | dave.mehringer | 2010-03-15 14:40:41 -0600 (Mon, 15 Mar 2010) | 14 lines
Changed paths:
  M /active/casacore/images/Images/ImageAnalysis.cc


      New Development: Yes/No
           JIRA Issue: No/Yes List JIRA ticket.
     Ready to Release: Yes/No
    Interface Changes: Yes/No
What Interface Changed: Please list interface changes
        Test Programs: List test programs
 Put in Release Notes: Yes/No
            Module(s): Module Names change impacts.
          Description: corrected LogOrigin call

	




git-svn-id: https://svn.cv.nrao.edu/svn/casa/test@10763 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/alma/implement/MS2ASDM/MS2ASDM.cc
@@ -719,7 +719,7 @@ namespace casa {
 	
 	unsigned long long timev = (unsigned long long) floor((time()(mainTabRow))*1E9); // what units? nanoseconds
 	unsigned long long intervalv = (unsigned long long) floor(interval()(mainTabRow)*1E9);
-	vector< unsigned long > flags;
+	vector< unsigned int > flags;
 	vector< long long > actualTimes;
 	vector< long long > actualDurations;
 	vector< float > zeroLags; // LAG_DATA, optional column, not used for the moment
@@ -824,7 +824,7 @@ namespace casa {
 	      }
 	    }
 	  }	
-	  unsigned long ul;
+	  unsigned int ul;
 	  for(uInt i=0; i<numSpectralPoint; i++){
 	    for(uInt j=0; j<numStokesMS; j++){
 	      if(skipCorr_p[PolId][j]){
@@ -867,7 +867,7 @@ namespace casa {
 			      autoData);         // single dish data.  
 
 	integrationNum++;
-	datasize += flags.size() * sizeof(unsigned long)
+	datasize += flags.size() * sizeof(unsigned int)
 	  + actualTimes.size() * sizeof( long long )
 	  + actualDurations.size() * sizeof( long long )
 	  + zeroLags.size() * sizeof( float )
